pyproject.toml
[tool.poetry]
name = "finance-system"
version = "0.1.0"
description = "Finance Dashboard Backend System"
authors = ["Mohammed Hussian"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
fastapi = "^0.115.0"
pydantic = "^2.8.0"
pydantic-settings = "^2.4.0"
uvicorn = { extras = ["standard"], version = "^0.30.0" }
sqlalchemy = "^2.0.0"
aiosqlite = "^0.20.0"
alembic = "^1.13.0"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-multipart = "^0.0.9"
greenlet = "^3.0.0"
bcrypt = "4.0.1"
slowapi = ">=0.1.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"
python-dotenv = "^1.0.0"
black = "^24.0.0"
isort = "^5.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
